+++ src/relay/backend/compile_engine.cc
@@ -21,11 +21,8 @@
  * \file relay/backend/compile_engine.cc
  * \brief Internal compialtion engine.
  */
-#include "compile_engine.h"
-
 #include <tvm/schedule.h>
 #include <tvm/packed_func_ext.h>
-#include <tvm/ir.h>
 #include <tvm/operation.h>
 #include <tvm/runtime/registry.h>
 #include <tvm/relay/attrs/device_copy.h>
@@ -39,9 +36,8 @@
 #include <functional>
 #include <vector>
 #include <unordered_map>
-
-#include "contrib/contrib_codegen.h"
 #include "../ir/type_functor.h"
+#include "compile_engine.h"
 
 namespace tvm {
 namespace relay {
@@ -598,15 +594,8 @@ class CompileEngineImpl : public CompileEngineNode {
   PackedFunc JIT(const CCacheKey& key) final {
     CCacheValue value = LowerInternal(key);
     if (value->packed_func != nullptr) return value->packed_func;
-    // Handle 3rd party generated code library.
-    if (value->lib.operator->()) {
-      auto name = FunctionGetAttr(key->source_func, attr::kFuncName);
-      const tvm::ir::StringImm* func_name = name.as<tvm::ir::StringImm>();
-      CHECK(func_name);
-      value->lib.GetFunction("init")();
-      value->packed_func = value->lib.GetFunction(func_name->value);
-    } else if (const auto* f = runtime::Registry::Get("relay.backend.build")) {
-      // build the function.
+    // build the function.
+    if (const auto* f = runtime::Registry::Get("relay.backend.build")) {
       tvm::runtime::Module m = (*f)(value->cached_func->funcs, key->target);
       value->packed_func = m.GetFunction(value->cached_func->func_name);
     } else {
@@ -659,22 +648,6 @@ class CompileEngineImpl : public CompileEngineNode {
       value->use_count = 0;
       cache_[key] = value;
     }
-
-    if (key->source_func->IsExternal()) {
-      auto compiler = FunctionGetAttr(key->source_func, attr::kExternal);
-      const tvm::ir::StringImm* code_gen = compiler.as<tvm::ir::StringImm>();
-      CHECK(code_gen);
-      std::string ext_name = "relay.ext." + code_gen->value;
-      auto pf = tvm::runtime::Registry::Get(ext_name);
-      CHECK(pf) << "Failed to find the codegen tool for " << ext_name << "\n";
-
-      // Invoke the 3rd party codegen to generate a library for the subgraph.
-      runtime::Module mod = (*pf)(key->source_func);
-      value->lib = mod;
-      value->cached_func = CachedFunc();
-      return value;
-    }
-
     // Enforce use the target.
     With<Target> target_scope(key->target);
