+++ src/relay/backend/compile_engine.h
@@ -137,8 +137,6 @@ class CCacheValueNode : public Node {
   CachedFunc cached_func;
   /*! \brief Result of Packed function generated by JIT */
   PackedFunc packed_func;
-  /*! \brief An external library generated by the 3rd party codegen. */
-  runtime::Module lib;
   /*! \brief usage statistics */
   int use_count{0};
